--- domain/src/main/java/team/aliens/domain/_param/auth/verify/CheckIdExistsResponse.kt
@@ -1,5 +1,0 @@
-package team.aliens.domain._param.auth.verify
-
-data class CheckIdExistsResponse(
-    val email: String,
-)
